// src/schemas/event.ts
import { z } from "zod";

export const createEventSchema = z.object({
  title: z.string().min(1, "Title is required"),
  description: z.string().optional(),
  date: z.date(),
  startTime: z
    .string()
    .regex(/^([01]\d|2[0-3]):([0-5]\d)$/, "Invalid time format"), // Must be "HH:mm"
  endTime: z
    .string()
    .regex(/^([01]\d|2[0-3]):([0-5]\d)$/, "Invalid time format"), // Must be "HH:mm"
  color: z.string(),
  participantEmails: z.array(z.string().email()).optional(),
});

export const updateEventSchema = z.object({
  eventId: z.string(),
  title: z.string().min(1, "Title is required"),
  description: z.string().optional(),
  date: z.date(),
  startTime: z
    .string()
    .regex(/^([01]\d|2[0-3]):([0-5]\d)$/, "Invalid time format"), // Must be "HH:mm"
  endTime: z
    .string()
    .regex(/^([01]\d|2[0-3]):([0-5]\d)$/, "Invalid time format"), // Must be "HH:mm"
  color: z.string(),
  participantEmails: z.array(z.string().email()).optional(),
});